Fix Japan1 & Japan2
Currently the regex greedily captures J1 instances and puts the rest in notes. For example J1234 becomes J1: J12 and 34. Instead, we want to capture instances of 1-2 digits and 3-5 digits after J without worrying about order.

However, because we are using a * with an | in the capture group, this will do a last match wins semantics within consecutive instances. For example. J1234 J3456 will match with J3456 and J1234 will be dropped. However, this should just be fixed I think.